refactored viz props context into shared
