+++ visualizations/podium/components/Podium.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { useProps } from "../context/VizPropsProvider";
+import { useProps } from "../../../shared/context/VizPropsProvider/VizPropsProvider"; 
 import { useNerdGraphQuery } from "../../../shared/hooks/useNerdGraphQuery/useNerdGraphQuery";
 import PodiumPlace from "./PodiumPlace";
 import { Stack } from "nr1";
